% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-beta.R
\name{ci.beta}
\alias{ci.beta}
\alias{beta-method}
\title{Confidence intervals based on the beta distribution}
\usage{
ci.beta(x, w, level, type = c("tcz", "cc"))
}
\arguments{
\item{x}{a vector of stratum-specific counts of events}

\item{w}{a vector of stratum-specific weights}

\item{level}{confidence level for the returned confidence interval}

\item{type}{type of modification for the beta confidence interval}
}
\value{
a vector with the lower and upper bound of the confidence interval.
The estimate of the directly standardised rate and the level of confidence are
returned as attributes to this vector.
}
\description{
Confidence intervals for directly standardized rates
based on the beta distribution.
}
\details{
Uses the methods proposed by Tiwari et al (2006) and further explored by
Ng et al (2008).
}
\references{
Ng, Filardo, & Zheng (2008).
'Confidence interval estimating procedures for
standardized incidence rates.' \emph{Computational Statistics and Data Analysis} \strong{52}: 3501-3516.

Tiwari, Clegg, & Zou (2006).
'Efficient interval estimation for age-adjusted cancer rates.'
\emph{Statistical Methods in Medical Research} \strong{15}: 547-569.
}
