% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gamma.R
\name{ci.gamma}
\alias{ci.gamma}
\alias{gamma-method}
\title{Confidence intervals using Gamma methods}
\usage{
ci.gamma(x, w, level, type = c("max", "midp", "mean", "minmaxavg", "tcz"),
  ...)
}
\arguments{
\item{x}{a vector of stratum-specific counts of events}

\item{w}{a vector of stratum-specific weights}

\item{level}{confidence level for the returned confidence interval}

\item{type}{type of modification for the gamma confidence interval}

\item{...}{passed to \code{asht::wspoissonTest} when type = \code{'midp'}}
}
\description{
Confidence intervals for directly standardized rates based on the gamma distribution.
}
\details{
Calls \code{asht::wspoissonTest} with the appropriate type modification for the gamma confidence interval (\code{wmtype}) and
\code{'mid.p'} value. Note that the \code{'mid.p'} method uses \code{wmtype = 'max'}.
}
\references{
Fay & Feuer (1997).
'Confidence intervals for directly standardized rates:
a method based on the gamma distribution.' \emph{Statistics in Medicine}. \strong{16}: 791-801.

Fay & Kim (2017).
'Confidence intervals for directly standardized rates
using mid-p gamma intervals.' \emph{Biometrical Journal}. doi:10.1002/bimj.201600111.

Ng, Filardo, & Zheng (2008).
'Confidence interval estimating procedures for
standardized incidence rates.' \emph{Computational Statistics and Data Analysis} \strong{52}: 3501-3516.

Tiwari, Clegg, & Zou (2006).
'Efficient interval estimation for age-adjusted cancer rates.'
\emph{Statistical Methods in Medical Research} \strong{15}: 547-569.
}
\seealso{
\code{\link[asht]{wspoissonTest}}
}
