% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsr.R
\name{dsr}
\alias{dsr}
\title{Calculate directly standardized rate}
\usage{
dsr(x, n, w, ci.method = c("asymptotic", "moments", "gamma", "beta",
  "bootstrap"), level = 0.95, mult = 1000, ...)
}
\arguments{
\item{x}{a vector of strata-specific counts}

\item{n}{a vector of strata-specific time bases for counts}

\item{w}{a vector of strata-specific weights (or standard populations)}

\item{ci.method}{method used to calculate the confidence interval. See details}

\item{level}{the confidence level required}

\item{mult}{a factor to multiply the estimate to give rates per \code{mult}}

\item{...}{Further arguments passed to the confidence interval function.}
}
\value{
\code{dsr} returns an object of class "\code{dsr}" . The function \code{confint} is used
to return a confidence interval using a specified method

An object of class "\code{dsr}" is a list containing the following components:
\tabular{ll}{
\code{estimate} \tab the estimate of the directly standardised rate \cr
\code{lower} \tab lower bound of the confidence interval \cr
\code{upper} \tab upper bound of the confidence interval \cr
\code{level} \tab the level of confidence \cr
\code{method} \tab method used to calcalate the confidence interval \cr
\code{call} \tab the matched call \cr
\code{mult} \tab The multiplicative factor
to scale the final estimate \cr
\code{strata} \tab number of strata or summands \cr
}
}
\description{
Calculate a directly standardised rate with confidence interval
based on a specified method.
}
\details{
Five groupds of methods can be specified using \code{'ci.method'}, with variations on each
depending on the method. The five groups are:

\tabular{ll}{
\code{'asymptotic'} \tab Using the normal approximation of the
MLE distribution (or transformed MLE distribution).
See \code{\link{ci.asymptotic}} for more details and
the currently implemented transformations. \cr
\code{'moments'} \tab Moment matching based on Dobson et al (1991).
A variety methods for constructing the confidence interval on the
unweighted sum of \code{x}. - See \code{\link{ci.moments}} for more details. \cr
\code{'gamma'} \tab Based on the gamma distribution (Fay & Feuer 1997).
See \code{\link{ci.gamma}} for more details and modifications implemented. \cr
\code{'beta'} \tab Based on the beta distribution as proposed by
Tiwari et al (2006). See \code{\link{ci.beta}} for details and
the modifications implemented. \cr
\code{'bootstrap'} \tab Appromiximate Bootstrap Confidence
proposed by Swift (1995). \cr
}
}
\references{
Dobson, AJ, Kuulasmaa, K, Eberle, E and Scherer, J (1991) 'Confidence intervals
for weighted sums of Poisson parameters', \emph{Statistics in Medicine}, \strong{10}: 457—462.

Swift, MB (1995). 'Simple confidence intervals for standardized rates based on
the approximate bootstrap method', \emph{Statistics in Medicine}, \strong{14}, 1875—1888.

Fay & Feuer (1997). 'Confidence intervals for directly standardized rates:
a method based on the gamma distribution.' \emph{Statistics in Medicine}. \strong{16}: 791-801.

Tiwari, Clegg, & Zou (2006). 'Efficient interval estimation for age-adjusted
cancer rates.'  \emph{Statistical Methods in Medical Research} \strong{15}: 547-569.

Ng, Filardo, & Zheng (2008). 'Confidence interval estimating procedures for
standardized incidence rates.' \emph{Computational Statistics and Data Analysis}
\strong{52} 3501-3516.
}
