% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-moments.R
\name{ci.moments}
\alias{ci.moments}
\alias{ci.dobson}
\title{Confidence intervals for Directly Standardised Rates using moment match methods}
\usage{
ci.moments(x, w, level, type = c("dobson", "boise.monson", "normal",
  "wilson.hilferty", "byar", "midp", "approx.midp", "simple.midp"))
}
\arguments{
\item{x}{a vector of stratum-specific counts of events}

\item{w}{a vector of stratum-specific weights}

\item{level}{confidence level for the returned confidence interval}

\item{type}{type of approximation for the poisson confidence interval of the unweighted sum}
}
\value{
a vector with the lower and upper bound of the confidence interval.
The estimate of the directly standardised rate and the level of confidence are
returned as attributes to this vector
}
\description{
Confidence intervals for directly standardized rates based on
the approximation proposed by Dobson, Kuulasmaa, Eberle and Scherer (1991).
In addition to the method proposed by Dobson et al, the various approximations
implemented by Ng, Filardo & Zheng (2008) are implemented.
}
\references{
Dobson, AJ, Kuulasmaa, K, Eberle, E and Scherer, J (1991)
'Confidence intervals for weighted sums of Poisson parameters',
\emph{Statistics in Medicine}, \strong{10}: 457—462.

Ng, Filardo, & Zheng (2008). 'Confidence interval estimating procedures for standardized incidence rates.'
\emph{Computational Statistics and Data Analysis} \strong{52}: 3501-3516.
}
